Add files via upload
This mod gives life to the eMylo YSAF033 bought from
https://www.amazon.fr/dp/B0C243M3PQ
Temperature and Humidity entities work, battery still need research. Clock and Date on display will update too.
This quirk was built for a HA 2023.1.7-20 installed on a Synology DS920+ as SynoCommunity Package, not docker or VM, might not work on last HA Version.